refactor: rename package to test to avoid conflicts
