+++ internal/kubernetes/job_test.go
@@ -11,7 +11,7 @@ import (
 
 	// internal packages
 	structInternal "volume-cleaner/internal/structure"
-	testInternal "volume-cleaner/internal/testing"
+	testInternal "volume-cleaner/internal/tests"
 )
 
 func TestIsStale(t *testing.T) {
